Read snapshot only for allowed tables
If table or database not included into list of tables or databases
allowed for replication they are not locked and shouldn't be read.

+++ debezium-connector-mysql/src/main/java/io/debezium/connector/mysql/SnapshotReader.java
@@ -462,6 +462,9 @@ public class SnapshotReader extends AbstractReader {
                             break;
                         }
                         String dbName = entry.getKey();
+                        if (!createTableFilters.databaseFilter().test(dbName)) {
+                            continue;
+                        }
                         // First drop, create, and then use the named database ...
                         schema.applyDdl(source, dbName, "DROP DATABASE IF EXISTS " + quote(dbName), this::enqueueSchemaChanges);
                         schema.applyDdl(source, dbName, "CREATE DATABASE " + quote(dbName), this::enqueueSchemaChanges);
@@ -469,6 +472,9 @@ public class SnapshotReader extends AbstractReader {
                         for (TableId tableId : entry.getValue()) {
                             if (!isRunning()) {
                                 break;
+                            }
+                            if (!createTableFilters.tableFilter().test(tableId)) {
+                                continue;
                             }
                             sql.set("SHOW CREATE TABLE " + quote(tableId));
                             mysql.query(sql.get(), rs -> {
